+++ packages/bench/suites/n-body/components/Position.ts
@@ -1,6 +1,6 @@
 import { Types, defineComponent } from "@bitecs/classic";
 
 export const Position = defineComponent({
-  x: Types.f32,
-  y: Types.f32,
+  x: Types.f64,
+  y: Types.f64,
 });
